Improve layout reading API
